requirements.txt
alpaca-py==0.6.1
pandas==1.5.2
pandas_ta==0.3.14b0
python-dotenv==1.0.0
schedule==1.1.0